Tests/LibWeb/Text/expected/Wasm/WebAssembly-instantiate-streaming.txt

WebAssembly.instantiateStreaming
Hello from wasm!!!!!!
WebAssembly.compileStreaming
Hello from wasm!!!!!!
Sanity check
